python/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
pygame
openpyxl